@@ -2,12 +2,12 @@ package com.steveperkins.fitnessjiffy.test;
 
 import com.steveperkins.fitnessjiffy.Application;
 import com.steveperkins.fitnessjiffy.service.ReportDataService;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.context.web.WebAppConfiguration;
 
 import javax.sql.DataSource;
@@ -16,7 +16,7 @@ import java.sql.Statement;
 import java.text.SimpleDateFormat;
 import java.util.concurrent.TimeUnit;
 
-@RunWith(SpringJUnit4ClassRunner.class)
+@ExtendWith(SpringExtension.class)
 @WebAppConfiguration
 @ContextConfiguration(classes = {Application.class})
 public abstract class AbstractTest {
@@ -30,15 +30,16 @@ public abstract class AbstractTest {
 
     final SimpleDateFormat simpleDateFormat = new SimpleDateFormat("yyyy-MM-dd");
 
-    @Before
+    @BeforeEach
     public void before() throws Exception {
-        final Connection connection = dataSource.getConnection();
-        final Statement statement = connection.createStatement();
-        statement.execute("DROP ALL OBJECTS");
-        statement.execute("RUNSCRIPT FROM 'classpath:/backup.sql'");
+        try (Connection connection = dataSource.getConnection();
+             Statement statement = connection.createStatement()) {
+            statement.execute("DROP ALL OBJECTS");
+            statement.execute("RUNSCRIPT FROM 'classpath:/backup.sql'");
+        }
     }
 
-    @After
+    @AfterEach
     public void after() throws InterruptedException {
         while (!reportDataService.isIdle()) {
             Thread.sleep(TimeUnit.SECONDS.toMillis(1));
